refactor(commands): split skill_cmds.rs (1853 lines) into submodules
U-16 (<=800 lines/file) was violated. Re-split the file along the same
lines as the earlier PR #124 attempt (which was lost during merge), now
including the additional cmd_save changes from #120 and rollback
surface from #136.

- skill_cmds/mod.rs        8
- add.rs                 131
- project.rs             260
- capture.rs             347
- save.rs                132
- import_observed.rs     319
- monitor.rs             538
- snapshot.rs             42
- shared.rs              212

All public App::cmd_* signatures unchanged. Also drop a redundant
JSON-roundtrip expect() in monitor.rs by using the original skill_id
String directly.
