+++ src/commands/skill_cmds/mod.rs
@@ -1,0 +1,8 @@
+mod add;
+mod capture;
+mod import_observed;
+mod monitor;
+mod project;
+mod save;
+mod shared;
+mod snapshot;
